rename editor -> editMode
